Don't turn error messages to int for debug
Have moved to not turning error messages into integers, because they are
easier to debug in textual form.

+++ src/main/java/no/ntnu/imt3281/ludo/common/MessageUtility.java
@@ -10,7 +10,7 @@ public class MessageUtility {
         var error = new JSONObject();
         error.put(FieldNames.ID, requestID);
         var codes = new JSONArray();
-        codes.put(code.toInt());
+        codes.put(code);
         error.put(FieldNames.CODE, codes);
         errors.put(error);
     }
